notification_service/requirements.txt
djangorestframework
django-environ
gunicorn
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
pika
requests